Update ruby docs to fix broken markdown
It appears that the intention of the docs was to have a separate section for error handling but got cut off between the comments and headers and code blocks.  

This PR attempts to fix the formatting and make it so that copying the code doesn't have errors in it.

Thanks!